+++ src/app/loadout/loadout-ui/LoadoutSubclassSection.tsx
@@ -9,6 +9,7 @@ import { AppIcon, powerActionIcon } from 'app/shell/icons';
 import { getDefaultAbilityChoiceHash, getSocketsByIndexes } from 'app/utils/socket-utils';
 import clsx from 'clsx';
 import { SocketCategoryHashes } from 'data/d2/generated-enums';
+import _ from 'lodash';
 import React, { useMemo } from 'react';
 import { createGetModRenderKey } from '../mod-utils';
 import EmptySubclass from './EmptySubclass';
@@ -19,9 +20,23 @@ export function getSubclassPlugs(
   defs: D2ManifestDefinitions,
   subclass: ResolvedLoadoutItem | undefined
 ) {
-  const plugs: PluggableInventoryItemDefinition[] = [];
+  const plugs: {
+    plug: PluggableInventoryItemDefinition;
+    active: boolean;
+    canBeEmptied: boolean;
+  }[] = [];
 
   if (subclass?.item.sockets?.categories) {
+    const aspects = subclass.item.sockets.categories.find(
+      (c) => c.category.hash === SocketCategoryHashes.Aspects
+    );
+    let activeFragments = _.sumBy(aspects?.socketIndexes, (aspectIndex) => {
+      const override = subclass.loadoutItem.socketOverrides?.[aspectIndex];
+      const def = override ? defs.InventoryItem.get(override) : undefined;
+
+      return def?.plug?.energyCapacity?.capacityValue || 0;
+    });
+
     for (const category of subclass.item.sockets.categories) {
       const showInitial =
         category.category.hash !== SocketCategoryHashes.Aspects &&
@@ -34,7 +49,9 @@ export function getSubclassPlugs(
         const hash = override || (showInitial && initial);
         const plug = hash && defs.InventoryItem.get(hash);
         if (plug && isPluggableItem(plug)) {
-          plugs.push(plug);
+          const active =
+            category.category.hash !== SocketCategoryHashes.Fragments || activeFragments-- > 0;
+          plugs.push({ plug, active, canBeEmptied: !showInitial });
         }
       }
     }
@@ -91,7 +108,12 @@ export default function LoadoutSubclassSection({
       {plugs.length ? (
         <div className={styles.subclassMods}>
           {plugs?.map((plug) => (
-            <PlugDef key={getModRenderKey(plug)} plug={plug} />
+            <PlugDef
+              className={clsx({ [styles.missingItem]: !plug.active })}
+              tooltipWarning={!plug.active ? t('Loadouts.InsufficientFragmentCapacity') : ''}
+              key={getModRenderKey(plug.plug)}
+              plug={plug.plug}
+            />
           ))}
         </div>
       ) : (